Results from discussions on PR and a bit more refactoring
