+++ app/build.gradle
@@ -41,8 +41,6 @@ dependencies {
     compile "com.squareup.retrofit2:adapter-rxjava2:2.3.0"
     compile "com.squareup.retrofit2:converter-gson:2.3.0"
     compile "io.reactivex.rxjava2:rxandroid:2.0.1"
-
-    //   compile 'android.arch.lifecycle:reactivestreams:' + rootProject.archLifecycleVersion
     compile 'android.arch.lifecycle:extensions:' + rootProject.archLifecycleVersion
     compile 'android.arch.persistence.room:runtime:' + rootProject.archRoomVersion
     compile 'android.arch.persistence.room:rxjava2:' + rootProject.archRoomVersion
@@ -51,7 +49,6 @@ dependencies {
     compile 'com.android.support:design:' + rootProject.supportLibVersion
     compile 'com.android.support:cardview-v7:' + rootProject.supportLibVersion
     compile 'org.jetbrains.kotlin:kotlin-stdlib-jre7:' + kotlin_version
-    //compile 'com.google.code.gson:gson:2.7'
 }
 repositories {
     mavenCentral()
