+++ app/src/main/kotlin/github/com/raccok/dota2androidapp/Utilities/Utilities.kt
@@ -2,6 +2,7 @@ package github.com.raccok.dota2androidapp.utilities
 
 import android.content.Context
 import android.content.pm.PackageManager
+import android.net.ConnectivityManager
 import android.support.v4.content.ContextCompat
 import android.widget.Toast
 
@@ -24,3 +25,8 @@ fun appIsMissingPermissions(context: Context): Boolean {
 
 private fun appHasPermission(context: Context, permission: String): Boolean =
         ContextCompat.checkSelfPermission(context, permission) == PackageManager.PERMISSION_GRANTED
+
+fun deviceIsOnline(connectivityMgr: ConnectivityManager?): Boolean {
+    val netInfo = connectivityMgr?.activeNetworkInfo
+    return netInfo != null && netInfo.isConnectedOrConnecting
+}
