+++ app/src/main/kotlin/github/com/raccok/dota2androidapp/sources/remote/Dota2OfficialAPIService.kt
@@ -1,6 +1,7 @@
 package github.com.raccok.dota2androidapp.sources.remote
 
 import github.com.raccok.dota2androidapp.api.HeroDataModelResponse
+import github.com.raccok.dota2androidapp.utilities.URLStrings
 import io.reactivex.Observable
 import retrofit2.Retrofit
 import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
@@ -8,21 +9,20 @@ import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-interface Dota2ApiService {
+interface Dota2OfficialAPIService {
     @GET("IEconDOTA2_570/GetHeroes/v1")
     fun fetchLocalizedHeroData(@Query("key") steamApiKey: String,
                                @Query("language") language: String): Observable<HeroDataModelResponse.Result>
 
     companion object {
-        //TODO: Put this in a helper class of constant strings.
-        private const val STEAM_API_URL = "http://api.steampowered.com/"
-
-        fun create(): Dota2ApiService {
+        fun create(): Dota2OfficialAPIService {
             val retrofit = Retrofit.Builder().addCallAdapterFactory(RxJava2CallAdapterFactory.create())
                     .addConverterFactory(GsonConverterFactory.create())
-                    .baseUrl(STEAM_API_URL)
+                    .baseUrl(URLStrings.STEAM_API)
                     .build()
-            return retrofit.create(Dota2ApiService::class.java)
+            return retrofit.create(Dota2OfficialAPIService::class.java)
         }
     }
 }
+
+val sDota2OfficialAPIService by lazy { Dota2OfficialAPIService.create() }
