+++ app/src/main/kotlin/github/com/raccok/dota2androidapp/utilities/PreferenceStrings.kt
@@ -1,6 +1,5 @@
-// TODO(Rhacco): maybe include these in strings.xml too?
 class PreferenceStrings {
     companion object {
-        const val PREF_FAV_HERO = "fav_hero"
+        const val FAV_HERO = "fav_hero"
     }
 }
