+++ app/src/main/kotlin/github/com/raccok/dota2androidapp/utilities/URLStrings.kt
@@ -1,0 +1,7 @@
+package github.com.raccok.dota2androidapp.utilities
+
+class URLStrings {
+    companion object {
+        const val STEAM_API = "http://api.steampowered.com/"
+    }
+}
